+++ Changes.md
@@ -1,11 +1,16 @@
 Version Information
 ===================
 
+Version 400.0.3
+---------------
+1. Add 'parenttheme' setting for dynamic parent theme support.  Initial supported parents: Moove and Boost.
+2. Add 'Musi' as a supported parent theme.
+
 Version 400.0.2
 ---------------
-  1. Add 'showhome' setting.
-  2. Add 'hvpcustomcss' setting.
+1. Add 'showhome' setting.
+2. Add 'hvpcustomcss' setting.
 
 Version 400.0.1
 ---------------
-  1. Initial version.
+1. Initial version.